% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcgaretriever_v_1_10_3.R
\name{check_TCGAretriever_connection}
\alias{check_TCGAretriever_connection}
\title{Check Connection to the cBioPortal Server}
\usage{
check_TCGAretriever_connection()
}
\value{
logical; `TRUE` if the system can connect and retrieve data 
within the time limit, `FALSE` otherwise.
}
\description{
Performs a connection check to determine if the remote cBioPortal data 
server is reachable and responsive.
}
\details{
This function attempts to fetch cancer studies within a strict 2-second 
time limit. If the function returns `FALSE`, it may indicate:
\itemize{
  \item The server is currently down for maintenance.
  \item The user's internet connection is inactive or blocked by a firewall.
  \item The connection speed is too slow to support reliable data retrieval 
        for larger datasets.
}
}
\examples{
\dontrun{
check_TCGAretriever_connection()
} 

}
\references{
\url{https://www.data-pulse.com/dev_site/TCGAretriever/}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
