% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcgaretriever_v_1_10_1.R
\name{get_cancer_studies}
\alias{get_cancer_studies}
\title{Retrieve the List of Cancer Studies Available at cbioportal.}
\usage{
get_cancer_studies(dryrun = FALSE)
}
\arguments{
\item{dryrun}{Logical. If TRUE, all other arguments (if any) are ignored and 
a representative example is returned as output. No Internet connection is 
required for executing the operation when `dryrun` is TRUE.}
}
\value{
Data Frame including cancer study information.
}
\description{
Retrieve information about the studies or datasets available at cbioportal.org. 
Information include a `studyId`, description, references, and more.
}
\examples{
# Set `dryrun = FALSE` (default option) in production!
all_studies <- get_cancer_studies(dryrun = TRUE)
utils::head(all_studies)



}
\references{
\url{https://www.data-pulse.com/dev_site/TCGAretriever/}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
