% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcgaretriever_v_1_10_1.R
\name{get_cancer_types}
\alias{get_cancer_types}
\title{Retrieve Cancer Types.}
\usage{
get_cancer_types(dryrun = FALSE)
}
\arguments{
\item{dryrun}{Logical. If TRUE, all other arguments (if any) are ignored and 
a representative example is returned as output. No Internet connection is 
required for executing the operation when `dryrun` is TRUE.}
}
\value{
A data.frame including cancer type information.
}
\description{
Retrieve information about cancer types and corresponding 
abbreviations from cbioportal.org. 
Information include identifiers, names, and parental cancer type.
}
\examples{
# Set `dryrun = FALSE` (default option) in production!
all_canc <- get_cancer_types(dryrun = TRUE)
utils::head(all_canc)

}
\references{
\url{https://www.data-pulse.com/dev_site/TCGAretriever/}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
