% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAretr_scr.R
\name{expand_cases}
\alias{expand_cases}
\title{Explode TCGA Case Identifiers from a TCGA Study}
\usage{
expand_cases(csid = NULL)
}
\arguments{
\item{csid}{string corresponding to a TCGA Cancer Study identifier}
}
\value{
list containing as many elements as TCGA case lists 
available for a given TCGA Study. Each element is a list containing two elements: 
\itemize{
  \item a string corresponding to the Id of the case list as defined by TCGA
  \item character vector including all case IDs corresponding to the case list
}
}
\description{
Each TCGA Study includes one or more "case lists". These are lists of 
sample/patient identifiers. All case lists of a study of interest are 
retrieved and the individual case identifiers are expanded and returned
}
\examples{
expand_cases("blca_tcga")

}
