% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcgaretriever_v_1_7.R
\name{get_cancer_studies}
\alias{get_cancer_studies}
\title{Retrieve a List of Cancer Studies Available at TCGA}
\usage{
get_cancer_studies()
}
\value{
Data Frame including one study per row and three columns.
}
\description{
Retrieve information about the different TCGA studies that are available at cBioPortal. 
Information include a cancer_study_id, a name of the study and a description for each study.
}
\examples{
all_studies <- get_cancer_studies()
message(paste("There are", nrow(all_studies), "studies currently available..."))

}
\references{
\url{https://www.data-pulse.com/dev_site/TCGAretriever/}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
