% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcgaretriever_v_1_7.R
\name{get_cancer_types}
\alias{get_cancer_types}
\title{Retrieve a List of Cancer Types as Defined by the TCGA Guidelines}
\usage{
get_cancer_types()
}
\value{
A data.frame with one row per cancer type and two columns
}
\description{
Retrieve information about the different types of cancer that may be 
included in TCGA Studies. Information include Identifier and Cancer Name.
}
\examples{
all_canc <- get_cancer_types()
message(paste("There are", nrow(all_canc), "cancer indications..."))

}
\references{
\url{https://www.data-pulse.com/dev_site/TCGAretriever/}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
