% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcgaretriever_v_1_7.R
\name{get_clinical_data}
\alias{get_clinical_data}
\title{Retrieve Clinical Information from a TCGA Study}
\usage{
get_clinical_data(case_id = "blca_tcga_all")
}
\arguments{
\item{case_id}{string corresponding to the case_list identifier of a specific list of cases of interest}
}
\value{
data.frame including one row per patient/case/sample
}
\description{
Retrieve Information about the Patients included in a TCGA Study of Interest. 
Each patient is associates with a case_id. Each case_id is accompained by a set 
of clinical information that may include sex, age, therapeutic regimen, Tumor Staging, 
vital status and others. NA are allowed.
}
\examples{
clinic_data <- get_clinical_data("blca_tcga_all")
clinic_data

}
\references{
\url{https://www.data-pulse.com/dev_site/TCGAretriever/}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
