% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcgaretriever_v_1_7.R
\name{get_genetic_profiles}
\alias{get_genetic_profiles}
\title{Retrieve Genetic Profiles for a TCGA Study of Interest}
\usage{
get_genetic_profiles(csid = NULL)
}
\arguments{
\item{csid}{string corresponding to the cancer study id of interest}
}
\value{
data.frame including one row per genetic profile and six columns
}
\description{
Retrieve Information about all genetic profiles associated with a TCGA Study of interest.
Each TCGA Study includes one or more kind of molecular analyses whose results are 
referred to as genetic profiles.
}
\examples{
get_genetic_profiles("blca_tcga")

}
\references{
\url{https://www.data-pulse.com/dev_site/TCGAretriever/}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
