\name{TCGAretriever-package}
\alias{TCGAretriever-package}
\alias{TCGAretriever}
\docType{package}
\title{Retrieve Genomic and Clinical Data from TCGA}
\description{The Cancer Genome Atlas (TCGA) is a program aimed at improving our understanding of Cancer Biology. Several TCGA Datasets are available online. TCGAretriever helps accessing and downloading TCGA data via the cBioPortal API. Features of TCGAretriever are: 1) it is simple and works as promised, unlike some other available packages; 2) it is tailored for downloading large volumes of data. }
\author{
Damiano Fantini <damiano.fantini@gmail.com>
}
\references{
Reference 1: http://www.biotechworld.it/bioinf/2016/07/11/tcga-data-via-tcgaretriever/
Reference 2: http://www.cbioportal.org/web_api.jsp
Reference 3: http://cancergenome.nih.gov/abouttcga/
}
