% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcgaretriever_v_1_7.R
\docType{package}
\name{TCGAretriever-package}
\alias{TCGAretriever}
\alias{TCGAretriever-package}
\title{Retrieve Genomic and Clinical Data from CBioPortal}
\description{
The Cancer Genome Atlas (TCGA) is a scientific and medical program 
aimed at improving our understanding of Cancer Biology. 
Several TCGA Datasets are available online, and some are hosted
on cBioPortal, which is an open-access, open-source resource for 
interactive exploration of multidimensional cancer genomics data sets. 
TCGAretriever helps accessing and downloading TCGA data via the 
cBioPortal API. Features of TCGAretriever are: 
\itemize{
  \item it is simple and reliable
  \item it is tailored for downloading large volumes of data
}
}
\references{
\itemize{
  \item \url{https://www.data-pulse.com/dev_site/TCGAretriever/}
  \item \url{http://www.cbioportal.org/} 
  \item \url{http://cancergenome.nih.gov/abouttcga/}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.data-pulse.com/dev_site/TCGAretriever/}
}

}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
\keyword{internal}
